proteinweb/requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.0.5
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.0.3
django-bootstrap4==2.2.0
djangorestframework==3.11.1
drf-writable-nested==0.6.3
factory-boy==3.0.1
Faker==9.8.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
msgpack==1.0.3
numpy==1.17.3
packaging==21.3
Pillow==7.2.0
prompt-toolkit==3.0.26
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
